Fully playable
Monsters and Zombies all spawn
Collissions implemented
Audio implemented

Still to do:

very little, maybe try to find some music, and a bit of graphical
polish for the menu?

I noticed a bug where the phone can actually go to sleep while you're
playing. There must be a way to stop it doing this, but it's 2:30am, so
I'll leave it for you ;)

I'll be in around lunchtime, see you then.